Added: Updated with api

+++ src/main/java/com/bridgelabz/bookstorebookservice/dto/BookDto.java
@@ -8,8 +8,8 @@ public class BookDto {
     private String author;
     private String description;
     private String logo;
-    private Long price;
-    private Long quantity;
+    private int price;
+    private int quantity;
 
 
 }
